locker/requirements.txt
peewee==3.17.9
python-dotenv==1.0.1
requests==2.32.3
